kraken/money-mode-router/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
psycopg[binary]>=3.1.18
